modules: don't add non-existant modules to autoload
Should solve issues where modules attempt to load the modules of their
dependencies, but the latter have not been generated.  Note that the
requirement here is that a module for the dependent DAG hash has to be
present, otherwise a warning is printed, and nothing will be
auto-loaded.

This could be potentially loosened to a version-match, at the risk of
loading incompatible modules (different variants could not be
discriminated reliably). Better safe than sorry.