Patches gene-expression-volume, py-atlalign and its dependent py-imgaug - Addresses naming issues raised by Gianluca
The main issue is that some packages <packages> referred to in the setup.py file
of <parent-package> are not visible for <parent-package> whereas the <packages>
python modules are actually installed.

The patch consists in removing the <packages> dependencies
from <parent-package>'s setup.py file (prunes install_requires).

The occurences of this problem are:
- parent-package = atlalign, packages = [tensorflow]
- parent-package = imgaug, packages = [opencv-python]
- parent-package = gene-expression-volume, packages = [tensorflow, lpips_tf]