deploy: consolidate stages
With the deployment of 2021, concretization of the specs was linear and
took a lot of time.  Thus the deployment of large teams was split into
multiple environments, e.g., sciene into science, ML, and DKE.

The new Spack of 2022 uses a "modern" conretizer, and using
environments, parallelizes on the environment level.  Due to GPFS
interacting with the Spack locking in an unreliable fashion, all
environments need to run subsequently.

Thus, to speed up our deployment, it would be beneficial to reduce the
number of environments we use:

* Concretization will be more parallel (more software concretized at the
  same time by Spack itself)
* Overhead from Spack preparation to run additional environments will be
  reduced