Patch pagebuffer check in `H5Fopen`.
The pagebuffer should be compatible with pHDF5, as long as
one doesn't use MPIO. This commit allows one request patching
HDF5 to weaken the check. The variant is called `+page_buffer_patch`.

This is only supported for HDF5 `v1.12.1`. It's used in the context
of reading morphologies from single-file HDF5 morphology containers.
One application that requires this is TD.