Fix arch spec on MacOS M2
The `apple-clang@14.0.0` compiler, which seems to be the default on
MacOS 12 with an M2 processor, does not support the `-mcpu=vortex` flag,
so `neuron` fails to build from source. According to this link:
https://community.arm.com/arm-community-blogs/b/tools-software-ides-blog/posts/compiler-flags-across-architectures-march-mtune-and-mcpu
it is advisable to always use `-mcpu=native` on ARM-based processors
when using both GNU and LLVM compilers.