deploy: update gmsh and boost variants on BB5
Build gmsh without support of useless libraries, including med which
depended on an older of hdf5. It induced specs of steps and dependent
packages like py-multiscale-run to duplicate a lot of packages.

Build boost with +serialization by default to meet highfive requirements and
thus reduce the number of boost packages installed.