deploy: set default versions when merging modules.
This came up with the recent Brayns release of 3.10, while a 3.8 version
was still deployed: the module system gave preference to the 3.8 version
due to lexical and not numerical sorting.

In the past we used special variables to pin the GCC and Python version
in the CI pipeline code itself.  This clearly does not scale, so move
the code into the merging of the modules (in Python!).

If there is more than one version of a module, first check if there is a
preference in the environment for a certain version (assumed to be
correct, existence of said version is not checked), use that. If not,
compare lexical and numerical sorting (via Spack's `Version`
functionality) and use the numerical version if not identical to the
best lexical one.